+++ MoralisDotNet/Moralis/Platform/Services/ClientServices/MoralisCloudFunctionService.cs
@@ -1,7 +1,5 @@
 ﻿using System;
 using System.Collections.Generic;
-using System.Threading;
-using Moralis.Platform.Utilities;
 using System.Threading.Tasks;
 using System.Net;
 using Moralis.Platform.Abstractions;
@@ -33,7 +31,7 @@ namespace Moralis.Platform.Services.ClientServices
         /// <returns>T - result of cloud function call.</returns>
         public async Task<T> CallFunctionAsync<T>(string name, IDictionary<string, object> parameters, string sessionToken, System.Threading.CancellationToken cancellationToken = default)
         {
-            MoralisCommand command = new MoralisCommand($"{ServerConnectionData.ParseEndpointCloudfunctions}/{Uri.EscapeDataString(name)}", method: "POST", sessionToken: sessionToken, data: parameters is { } && parameters.Count > 0 ? JsonSerializer.Serialize(parameters) : "{}");
+            MoralisCommand command = new MoralisCommand($"{ServerConnectionData.ParseEndpointBase}/{ServerConnectionData.ParseEndpointCloudfunctions}/{Uri.EscapeDataString(name)}", method: "POST", sessionToken: sessionToken, data: parameters is { } && parameters.Count > 0 ? JsonSerializer.Serialize(parameters) : "{}");
 
             Tuple<HttpStatusCode, string> cmdResult = await CommandRunner.RunCommandAsync(command, cancellationToken: cancellationToken);